better better structure of new  bsps
