+++ bsp/boards/nrf51822_crazyflie2/cf_crazyflie.c
@@ -1,69 +1,35 @@
 /**
+ *
  * @file cf_crazyflie.c
- * @brief crazyflie operations
- * @author Lan HUANG(YelloooBlue@outlook.com)
+ * @brief crazyflie basic operation, like boot
+ * @author Lan HUANG (YelloooBlue@outlook.com)
  * @date June 2024
  *
  */
 
-#include "stdint.h"
-#include "stdio.h"
-#include "string.h"
+
+#include <stdint.h>
+#include <stdio.h>
+#include <string.h>
 
 // crazyflie bsp
+#include "cf_crazyflie.h"
+#include "cf_platform.h"
 #include "cf_pm.h"
 #include "cf_systick.h"
-#include "cf_uart.h"
-#include "cf_syslink.h"
-#include "cf_platform.h"
-#include "cf_crazyflie.h"
 #include "cf_memory.h"
 #include "cf_ownet.h"
+#include "cf_uart.h"
+#include "cf_syslink.h"
+#include "cf_ctrp.h"
 
 //=========================== defines =========================================
 
-#define BLE
 #ifdef BLE
 int volatile bleEnabled = 1;
 #else
 int volatile bleEnabled = 0;
 #endif
-
-// Parameters index
-#define index_deck_bcFlow2 10
-#define index_commander_enHighLevel 89
-#define index_kalman_resetEstimation 112
-#define index_stabilizer_controller 140
-
-// Syslink & CRTP defines
-typedef enum
-{
-  CRTP_PORT_CONSOLE = 0x00,
-  CRTP_PORT_PARAM = 0x02,
-  CRTP_PORT_SETPOINT = 0x03,
-  CRTP_PORT_MEM = 0x04,
-  CRTP_PORT_LOG = 0x05,
-  CRTP_PORT_LOCALIZATION = 0x06,
-  CRTP_PORT_SETPOINT_GENERIC = 0x07,
-  CRTP_PORT_SETPOINT_HL = 0x08,
-  CRTP_PORT_PLATFORM = 0x0D,
-  CRTP_PORT_LINK = 0x0F,
-} CRTPPort;
-
-enum TrajectoryCommand_e
-{
-  COMMAND_SET_GROUP_MASK = 0, // Deprecated (removed after Dec 2024), use parameter hlCommander.groupmask instead
-  COMMAND_TAKEOFF = 1,        // Deprecated (removed after August 2023), use COMMAND_TAKEOFF_2
-  COMMAND_LAND = 2,           // Deprecated (removed after August 2023), use COMMAND_LAND_2
-  COMMAND_STOP = 3,
-  COMMAND_GO_TO = 4,
-  COMMAND_START_TRAJECTORY = 5,
-  COMMAND_DEFINE_TRAJECTORY = 6,
-  COMMAND_TAKEOFF_2 = 7,
-  COMMAND_LAND_2 = 8,
-  COMMAND_TAKEOFF_WITH_VELOCITY = 9,
-  COMMAND_LAND_WITH_VELOCITY = 10,
-};
 
 //=========================== variables =======================================
 static bool stm_ready = false;
@@ -74,15 +40,11 @@ static struct syslinkPacket slRxPacket;
 static struct syslinkPacket slTxPacket;
 
 //=========================== prototypes ======================================
-void _CTRPSend(void *data, uint8_t size, uint8_t port, uint8_t channel);
-void _CTRPSend_NULL();
-void _high_level_sendCommand(const enum TrajectoryCommand_e command_type, void *values, uint8_t size);
 
 //=========================== public ==========================================
 
 void crazyflieInit()
 {
-  // Crazyflie init
   pmInit();
   uartInit();
   systickInit();
@@ -260,7 +222,7 @@ void syslinkHandle()
   if (systickGetTick() - lastSyslinkSendTime > SYSLINK_SEND_PERIOD_MS)
   {
     lastSyslinkSendTime = systickGetTick();
-    _CTRPSend_NULL(); // just send NULL to keep syslink alive, otherwise the STM32 will blocked until a packet is received
+    CTRPSend_NULL(); // just send NULL to keep syslink alive, otherwise the STM32 will blocked until a packet is received
   }
 
   /*
@@ -269,149 +231,17 @@ void syslinkHandle()
     2. You can not send data via syslink too frequently, otherwise the NRF51822 will overflow (the LED_M3 will blink slowly).
     3. The OneWire Packet will be exchanged at the beginning, the decks on Crazyflie will be detected and initialized. [important!]
     4. It's best not to send too many packets at the beginning, otherwise the OneWire Packet might be lost.
-    
+
     @Author Lan HUANG(YelloooBlue@outlook.com) June 2024
   */
-
 }
 
-
-// Safety
 void EmergencyStop()
 {
   uint8_t data[1] = {0x03}; // 0x03:EMERGENCY_STOP
-  _CTRPSend(data, 1, CRTP_PORT_LOCALIZATION, 1);
-}
-
-
-
-// Parameters Operation
-void param_set(uint16_t id, const void *value, uint8_t len)
-{
-
-  uint8_t data[CRTP_MAX_DATA_SIZE];
-  data[0] = id & 0xFF;
-  data[1] = (id >> 8) & 0xFF;
-
-  memcpy(data + 2, value, len);
-
-  _CTRPSend(data, len + 2, CRTP_PORT_PARAM, 2);
-}
-
-// High-Level Commander
-// #@#@#@# After sending a high-level command, the param 'enHighLevel' should be set to 1 !!!!!!
-void high_level_enable()
-{
-  uint8_t data[1] = {1};
-  param_set(index_commander_enHighLevel, data, 1);
-}
-
-// Trajectory
-void high_level_takeoff(float absolute_height_m, float duration_s, float yaw)
-{
-
-  high_level_enable(); // let the high level commander be enabled (necessary)
-
-  struct data_takeoff_2 takeoffValues;
-  takeoffValues.groupMask = 0x00;
-
-  takeoffValues.height = absolute_height_m;
-  takeoffValues.duration = duration_s;
-
-  if (yaw == 0.0)
-  {
-    takeoffValues.yaw = 0.0;
-    takeoffValues.useCurrentYaw = true;
-  }
-  else
-  {
-    takeoffValues.yaw = yaw;
-    takeoffValues.useCurrentYaw = false;
-  }
-
-  _high_level_sendCommand(COMMAND_TAKEOFF_2, &takeoffValues, sizeof(takeoffValues));
-}
-
-void high_level_land(float absolute_height_m, float duration_s, float yaw)
-{
-
-  struct data_land_2 landValues;
-  landValues.groupMask = 0x00;
-
-  landValues.height = absolute_height_m;
-  landValues.duration = duration_s;
-
-  if (yaw == 0.0)
-  {
-    landValues.yaw = 0.0;
-    landValues.useCurrentYaw = true;
-  }
-  else
-  {
-    landValues.yaw = yaw;
-    landValues.useCurrentYaw = false;
-  }
-
-  _high_level_sendCommand(COMMAND_LAND_2, &landValues, sizeof(landValues));
-}
-
-void high_level_stop()
-{
-  struct data_stop stopValues;
-  stopValues.groupMask = 0x00;
-
-  _high_level_sendCommand(COMMAND_STOP, &stopValues, sizeof(stopValues));
-}
-
-void high_level_goto(float x, float y, float z, float yaw, float duration_s, bool relative)
-{
-  struct data_go_to go_toValues;
-  go_toValues.groupMask = 0x00;
-  go_toValues.relative = relative;
-  go_toValues.x = x;
-  go_toValues.y = y;
-  go_toValues.z = z;
-  go_toValues.yaw = yaw;
-  go_toValues.duration = duration_s;
-
-  _high_level_sendCommand(COMMAND_GO_TO, &go_toValues, sizeof(go_toValues));
+  CTRPSend(data, 1, CRTP_PORT_LOCALIZATION, 1);
 }
 
 //=========================== private =========================================
-
-// CTRP
-
-void _CTRPSend(void *data, uint8_t size, uint8_t port, uint8_t channel)
-{
-  struct CRTPPacket ctrp_pk;
-  ctrp_pk.port = port;
-  ctrp_pk.channel = channel;
-  ctrp_pk.reserved = 0;
-  ctrp_pk.size = size;
-  memcpy(ctrp_pk.data, data, size);
-
-  struct syslinkPacket syslink_pk;
-  syslink_pk.type = 0x00;
-  syslink_pk.length = ctrp_pk.size + 1;
-  memcpy(syslink_pk.data, &ctrp_pk.raw, syslink_pk.length); // syslink_data_len = ctrp_raw_len = ctrp_data_len + ctrp_header_len
-
-  syslinkSend(&syslink_pk);
-}
-
-// Send NULL to keep syslink alive. Port 15, channel 3
-void _CTRPSend_NULL()
-{
-  _CTRPSend(NULL, 0, CRTP_PORT_LINK, 3);
-}
-
-void _high_level_sendCommand(const enum TrajectoryCommand_e command_type, void *values, uint8_t size)
-{
-  uint8_t data[CRTP_MAX_DATA_SIZE];
-  data[0] = command_type;
-  memcpy(data + 1, values, size);
-  data[size + 1] = 0; // TODO:这里去掉0的话飞机就会翻车，原因未知（已确定不是数据拷贝长度的问题。将data数组初始化为0也是一个选择
-
-  _CTRPSend(data, size + 1, CRTP_PORT_SETPOINT_HL, 0); // ctrp_data_len = values_len + command_type_len
-}
 
 //=========================== interrup handlers ===============================
