+++ bsp/boards/nrf51822_crazyflie2/crazyflie/cf_param.h
@@ -1,0 +1,31 @@
+/**
+ *
+ * @file cf_param.h
+ * @brief Crazyflie Parameter Operation
+ *    this API is migrated from the higher level code like the python client
+ * @author Lan HUANG (YelloooBlue@outlook.com)
+ * @date June 2024
+ *
+ */
+
+#ifndef __CF_PARAM_H__
+#define __CF_PARAM_H__
+
+#include <stdint.h>
+
+//=========================== define ==========================================
+
+#define index_deck_bcFlow2 10
+#define index_commander_enHighLevel 89
+#define index_kalman_resetEstimation 112
+#define index_stabilizer_controller 140
+
+//=========================== typedef =========================================
+
+//=========================== variables =======================================
+
+//=========================== prototypes ======================================
+
+void param_set(uint16_t id, const void *value, uint8_t len);
+
+#endif // __CF_PARAM_H__
